requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.2
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-htmx==1.15.0
django-prose==1.2.2
Pillow==10.0.0
pytils==0.4.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
webencodings==0.5.1
